Track (previously deferred) emitted symbol; put back as deferred if unloaded.
`EmittedDeferredDecls` is a map from "global value emitted because a deferred definition was needed" to the entries of the DeferredDecls to be put back when the global value is unloaded.
Fix the unloading of definitions of emitted deferred symbols, and the unloading of declarations from the DeferredDecls.
forgetDecl() only needs to act on weak declarations with a definition (strong symbols are emitted, those without a definition cannot enter DeferredDecls).