Autoload subdirectory modulemaps with specific LangOpts
This patch already landed in https://reviews.llvm.org/rL336660 in Clang.

This was biting us to enable runtime modules in CMSSW.

Detailed desciption:
https://bugs.llvm.org/show_bug.cgi?id=37878

lookupModule was falling back to loadSubdirectoryModuleMaps when it couldn't
find ModuleName in (proper) search paths. This was causing iteration over all
files in the search path subdirectories for example "/usr/include/foobar" in
bugzilla case.

Users don't expect Clang to load modulemaps in subdirectories implicitly, and
also the disk access is not cheap.

if (AllowExtraModuleMapSearch) true with ObjC with @import ModuleName.